fix(vcs): address PR review — safety, timeouts, and code quality
- fix: never move shared branches (main) to arbitrary changes in
  suggestPushBookmarkForChange — always generate feature branch names
- fix: add 30s timeout to adhoc agent via Promise.race to prevent
  indefinite IPC hangs during PR draft generation
- fix: preserve user's 'auto' push bookmark choice across bookmark
  reloads by tracking activePushBookmarkInitialized flag
- fix: read fresh store state in handlePushAs to avoid race with stale
  render snapshot of bookmarks
- fix: validate source branch is pushed before gh pr create to surface
  clear error instead of confusing GitHub API failure
- fix: show error feedback in UI when PR draft generation fails
  (was silently swallowed with console.warn only)
- refactor: extract branch naming utils (inferConventionalType,
  slugifyBranchLabel) into electron/vcs/branch-naming.ts
- refactor: extract PR prompt/parse logic (buildPrDraftPrompt,
  parseDraft) into electron/agents/pr-draft.ts
- chore: add console.warn/debug to all silent catch blocks for
  debuggability
- chore: document SSH TOFU threat model in jj-runner.ts
